feat: renamed handlers folder to features
