+++ backend/UserService/UserService.Api/Startup.cs
@@ -3,8 +3,8 @@ using UserService.Api.Extensions;
 using UserService.Api.Filters;
 using UserService.Api.Middlewares;
 using UserService.Application.Extensions;
-using UserService.Application.Handlers.Email;
-using UserService.Application.Handlers.Jwt;
+using UserService.Application.Features.Email;
+using UserService.Application.Features.Jwt;
 using UserService.Application.Mapper;
 using UserService.DataAccess.Database.UnitOfWork;
 using UserService.DataAccess.Extensions;
