+++ backend/UserService/UserService.Application/Extensions/ServiceCollectionExtension.cs
@@ -3,9 +3,9 @@ using Microsoft.Extensions.DependencyInjection;
 using UserService.Application.Dto.EmailDtos;
 using UserService.Application.Dto.RoleDtos;
 using UserService.Application.Dto.UserDtos;
-using UserService.Application.Handlers.Jwt;
-using UserService.Application.Handlers.JwtUtilities;
-using UserService.Application.Handlers.Providers;
+using UserService.Application.Features.Jwt;
+using UserService.Application.Features.PasswordHasher;
+using UserService.Application.Features.Providers;
 using UserService.Application.Interfaces.Auth;
 using UserService.Application.Interfaces.Providers;
 using UserService.Application.Interfaces.Services;
